np.float64() and Python's builtin float give the same number, only their representation is different, so I got rid of all the np.float64().

+++ python/W6_main.py
@@ -9,11 +9,11 @@ if __name__ == '__main__':
     # don't need them in the user input part
     # SET THE CONSTANTS
     dtosec = 86400
-    big_G = np.float64(6.67259e-11)
-    Rjup = np.float64(7.15e7)
-    Rsun = np.float64(6.96e8)
-    Mjup = np.float64(1.9e27)
-    Msun = np.float64(1.99e30)
+    big_G = 6.67259e-11
+    Rjup = 7.15e7
+    Rsun = 6.96e8
+    Mjup = 1.9e27
+    Msun = 1.99e30
     HST_second = 5781.6
     HST_period = 0.06691666
 
@@ -22,11 +22,11 @@ if __name__ == '__main__':
     dataDir = os.path.join(mainDir, 'data')
 
     dtosec = 86400
-    big_G = np.float64(6.67259e-11)
-    Rjup = np.float64(7.15e7)
-    Rsun = np.float64(6.96e8)
-    Mjup = np.float64(1.9e27)
-    Msun = np.float64(1.99e30)
+    big_G = 6.67259e-11
+    Rjup = 7.15e7
+    Rsun = 6.96e8
+    Mjup = 1.9e27
+    Msun = 1.99e30
     HST_second = 5781.6
     HST_period = 0.06691666
 
@@ -38,16 +38,16 @@ if __name__ == '__main__':
     # SET-UP the parameters for the subroutine
     # ---------------------
     # PLANET PARAMETERS
-    rl = np.float64(0.13169232)  # Rp/R* estimate
-    epoch = np.float64(57879.633617736)  # in MJD
-    inclin = np.float64(88.41)  # this is converted into radians in the subroutine
+    rl = 0.13169232  # Rp/R* estimate
+    epoch = 57879.633617736  # in MJD
+    inclin = 88.41  # this is converted into radians in the subroutine
     ecc = 0.0  # set to zero and not used when circular
     omega = 0.0  # set to zero and not used when circular
-    Per = np.float64(3.36100239)  # in days, converted to seconds in subroutine
+    Per = 3.36100239  # in days, converted to seconds in subroutine
 
     persec = Per * dtosec
-    aor = np.float64(10.94)  # a/r* converted to system density for the subroutine
-    constant1 = (big_G * persec * persec / np.float32(4 * 3.1415927 * 3.1415927)) ** (1 / 3.)
+    aor = 10.94  # a/r* converted to system density for the subroutine
+    constant1 = (big_G * persec * persec / (4 * 3.1415927 * 3.1415927)) ** (1 / 3.)
     MsMpR = (aor / (constant1)) ** 3
 
     ld_model = '1D'   # Which limb darkening models to use, '1D' or '3D'
